+++ test/OpenTelemetry.Resources.Process.Tests/ProcessDetectorTests.cs
@@ -15,7 +15,6 @@ public class ProcessDetectorTests
         var resourceAttributes = resource.Attributes.ToDictionary(x => x.Key, x => x.Value);
 
         Assert.IsType<long>(resourceAttributes[ProcessSemanticConventions.AttributeProcessArgsCount]);
-        Assert.IsType<string>(resourceAttributes[ProcessSemanticConventions.AttributeProcessCommand]);
         Assert.IsType<string>(resourceAttributes[ProcessSemanticConventions.AttributeProcessCreationTime]);
         Assert.IsType<string>(resourceAttributes[ProcessSemanticConventions.AttributeProcessExecutablePath]);
         Assert.IsType<string>(resourceAttributes[ProcessSemanticConventions.AttributeProcessOwner]);