+++ src/OpenTelemetry.Resources.Process/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+* Extended `ProcessDetector` to include `process.executable.path`,
+  `process.working_directory`, `process.args_count`,`process.creation.time`,
+  `process.executable.name`, `process.interactive` as well as `process.title`
+  as per the latest OpenTelemetry Semantic Convention resource/entity definition.
+  ([#2971](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3347))
+
 ## 1.15.0-beta.1
 
 Released 2026-Jan-21