+++ src/OpenTelemetry.Resources.Process/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+* Add new process resource attributes: `process.args_count`,
+  `process.creation.time`, `process.executable.path`,
+  `process.title` and `process.working_directory`.
+  ([#4036](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4036))
+
 ## 1.15.1-beta.1
 
 Released 2026-Apr-21